+++ src/components/table/table.js
@@ -1,5 +1,6 @@
 import React from 'react';
 import classNames from 'classnames';
+import Immutable from 'immutable';
 import TableRow from './table-row';
 import TableCell from './table-cell';
 import TableHeader from './table-header';
@@ -76,6 +77,14 @@ class Table extends React.Component {
 
   static propTypes = {
     /**
+     * Data used to filter the data
+     *
+     * @property filter
+     * @type {Object}
+     */
+    filter: React.PropTypes.object,
+
+    /**
      * Emitted when table component changes e.g.
      * Pager, sorting, filter
      *
@@ -147,6 +156,20 @@ class Table extends React.Component {
    */
   componentDidMount() {
     this.resizeTable();
+  }
+
+  /**
+   * Lifecycle for after a update has happened
+   * If filter has changed then emit the on change event.
+   *
+   * @method componentWillReceiveProps
+   * @return {Void}
+   */
+  componentWillReceiveProps(nextProps) {
+    // if filter has changed, update the data
+    if (!Immutable.is(this.props.filter, nextProps.filter)) {
+      this.emitOnChangeCallback('filter', this.emitOptions(nextProps));
+    }
   }
 
   /**
@@ -238,7 +261,7 @@ class Table extends React.Component {
    * @return {Void}
    */
   onPagination = (currentPage, pageSize) => {
-    let options = this.emitOptions;
+    let options = this.emitOptions();
     options.currentPage = currentPage;
     options.pageSize = pageSize;
     this.emitOnChangeCallback('pager', options);
@@ -250,11 +273,12 @@ class Table extends React.Component {
    * @method emitOptions
    * @return {Object} options to emit
    */
-  get emitOptions() {
+  emitOptions = (props = this.props) => {
     return {
       // What if paginate if false - think about when next change functionality is added
-      currentPage: this.props.currentPage,
-      pageSize: this.props.pageSize
+      currentPage: props.currentPage,
+      filter: props.filter ? props.filter.toJS() : {},
+      pageSize: props.pageSize
     };
   }