+++ CHANGELOG.md
@@ -3,11 +3,13 @@
 ## New Components
 
 * Pager
+* Filter
 * Table Ajax
 
 ## Bug Fixes
 
 * TableCell and TableHeader can receive additional props.
+* Inputs no longer render a label if the input has no name or label props.
 
 # 0.6.0